# editor/server/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "editor API",
    "version": "0.1.0",
    "description": "Jinja2 テンプレート GUI 編集アプリのバックエンド API。\n\n`@editor/shared` の Repository 契約(Template / Auth / User / History / Part)を REST として設計したもの。一部は Phase 2 の設計(ハンドラ未実装)。"
  },
  "servers": [
    {
      "url": "/api",
      "description": "同一オリジン (server が SPA も配信)"
    }
  ],
...
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/build/project": {
      "post": {
        "tags": [
          "vivliostyle"
        ],
        "summary": "vivliostyle プロジェクト(zip)から PDF を生成",
        "operationId": "buildProject",
        "description": "リクエストボディに project zip を `application/zip` で送信する。**受理は許可リスト方式**で、1 エントリでも条件を外れた時点で zip 全体が 400 になる(該当エントリだけ無視する挙動ではない)。\n\n**展開して残せる拡張子**(小文字化した basename の末尾一致): `.html` `.htm` `.xhtml` `.xht` `.md` `.markdown` `.css` `.css.map` `.png` `.jpg` `.jpeg` `.svg` `.gif` `.webp` `.apng` `.ttf` `.otf` `.woff` `.woff2` `.json`。これ以外(`.js` `.mjs` `.cjs` `.ts` を含む)は 400。設定を `vivliostyle.config.js` 等で送っても「無視される」のではなく**拒否される**。\n\n**黙って捨てる OS ノイズ**(400 にはしない): `__MACOSX/**`・`.DS_Store`・`Thumbs.db`・`desktop.ini`・`._` で始まる名前。\n\n**拒否する npm 成果物**: `node_modules/` を含むパス・`package.json`・`package-lock.json`・`.npmrc`。\n\n**config は `vivliostyle.config.json` ちょうど 1 件**(名前は大小文字を無視して照合)。2 件以上あれば 400。無ければ単一エントリを既定として使う。内容は**厳密な JSON** のみで、コメント・末尾カンマ(JSONC)や配列(複数タスク)は 400。\n\n**受理するトップレベルキー**: `entry` `entryContext` `theme` `cover` `toc` `title` `author` `language` `readingProgression` `size` `copyAsset`。これ以外はキー名を挙げて 400 になる(`output` `workspaceDir` `vfm` `image` `timeout` `viewer` `server` `static` `pdfPostprocess` など vivliostyle CLI の正規フィールドも含む)。`base` は専用メッセージで 400 — 文書の配信先はサーバが固定する。\n\n**`theme` の制約**: 展開ルート内に実在する `.css` ファイルへの相対パスのみ。拡張子の判定は**大小文字を区別**する(`.CSS` は不可)。npm パッケージ指定子・`{specifier, import}` 形は不可。`entry` 要素ごとの `theme` にも同じ制約が掛かる。\n\n**400 応答**はすべて `kind=validation` で、`message` は日本語 1 文。拡張子違反は「取り込めるのは <許可拡張子一覧> だけです」、キー違反は「vivliostyle.config.json に指定できないキーがあります: <キー名>」のように**違反した対象を名指しする**(どのファイル・どのキーで落ちたかが応答から判る)。展開時の資源上限(ファイル数・展開後サイズ・圧縮比・目録サイズ)超過も 400 で、上限値を message に含める。413 になるのは**リクエスト本文そのもの**がサーバの `bodyLimit` を超えた場合だけ(fastify が返す)。\n\n任意クエリ: `entry`, `size`, `singleDoc`(`singleDoc` は zip に config が無いときだけ効き、config があるときは無視する)。\n\n**外部参照は拒否する**(`code=DOCUMENT_EXTERNAL_REF` の 400)。PDF は CSP の無い headless ブラウザで組版されるため、CSS からの取得はそのままビルドサーバの位置からの GET になる。検査対象は (1) リクエストの `css`、(2) HTML 中の `<style>` ブロック、(3) HTML の `style=\"…\"` 属性、(4) HTML の**取得系属性**(`<link href>` `<script src>` `<img src|srcset>` `<iframe src>` `<object data>` SVG の `href`/`xlink:href` 等。`<a href>` は組版中に取得を起こさないので対象外)、(zip 経路は加えて展開した `.css` / `.html` 全件)。拒否するのは **許可リスト外の at-rule**(`@import` `@use` 等。`@media` `@page` とマージンボックス `@bottom-center` 等は許可)と、**外部を指す URL 値**(scheme 付き・scheme 相対 `//host/x`・許可外の `data:`)で、`url()` に限らず`image-set(\"http://…\")` のような引用符文字列も含む。許可する `data:` は `data:image/png` `data:image/jpeg` `data:image/jpg` `data:image/gif` `data:image/webp` `data:font/` `data:application/font-woff` の接頭辞のみ(`data:image/svg+xml` は不可)。**相対 URL と断片(`#id`)は通る。むしろ必須である** — 文書は作業フォルダの `doc/` に置かれたものとして扱い、テンプレ CSS・共通フォント・テンプレ JS・画像を `../css/…` `../css/fonts/…` `../js/…` `../images/…` の相対パスで参照する。サーバは参照された実体だけを `doc/` の兄弟(`css/` `js/` `images/`)へ同梱する。文書直下基準の `css/…` `images/…` は `doc/css/…` を指し、同梱されない。画像は `../images/<名前>` と `../images/<フォルダ>/<名前>`(1 段まで)を同梱し、2 段以上は同梱しない。展開した CSS ファイルは置かれた位置を基準に相対参照を解決する。相対参照は `url()` で書くこと(引用符文字列の相対参照は解決されない)。同梱の実体が無い相対参照の `<link>` / `<script src>` は 400 にはせず要素ごと落とす(404 は組版のページ分割を止めるため)。**削らずに拒む**のは外部参照の方で、違反が 1 件でも PDF は生成されない。また、タグ境界が一意に決まらない HTML(閉じないタグ・コメント・`<style>`/`<script>`)は`code=DOCUMENT_UNPARSABLE` の 400 で拒む — 検査できない入力を通すとそれ自体が回避路になる。\n\n**文書内の JavaScript**: 組版時に実行される。組版エンジンは文書を再パースして script をビューアの window へ作り直すため、`<head>` で `DOMContentLoaded` に登録した処理は発火しない(実測。処理は body 末尾に置くこと)。実行されるのは HTML に直接書いたインライン `<script>` だけである。zip へ `.js` は同梱できず、`<script src>` は相対 URL の解決基準がずれて 404 になる(サーバはインライン展開しない)。なお組版ブラウザの **HTTP/HTTPS 通信**は、そのビルド専用の loopback オリジン 1 つだけへ中継される(それ以外は宛先が loopback でも 502 で落ちる)。ただし遮断の実体は HTTP プロキシ 1 本なので、**HTTP 以外の経路**(WebRTC の UDP 等)はこの中継を通らない — そこは残余リスクとして残る。\n\n",
        "requestBody": {
          "content": {
            "application/zip": {
              "schema": {
                "type": "string",
                "format": "binary",
                "description": "PDF バイナリ"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PDF バイナリ",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "description": "PDF バイナリ"
                }
              }
            }
          },
          "400": {
            "description": "リクエスト検証エラー (kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "403": {
            "description": "編集者権限が必要 (viewer は不可。kind=forbidden)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "413": {
            "description": "プロジェクトが大きすぎる (kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "500": {
            "description": "サーバ内部エラー (kind=unexpected)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
    "/build/merge": {
      "post": {
        "tags": [
          "vivliostyle"
        ],
        "summary": "複数のレンダリング済み文書を 1 つの PDF へ結合",
        "operationId": "buildMerge",
        "description": "`documents` の配列順に連結し、全体で通しページ番号を振った 1 つの PDF を返す。\n\n**外部参照は拒否する**(`code=DOCUMENT_EXTERNAL_REF` の 400)。PDF は CSP の無い headless ブラウザで組版されるため、CSS からの取得はそのままビルドサーバの位置からの GET になる。検査対象は (1) リクエストの `css`、(2) HTML 中の `<style>` ブロック、(3) HTML の `style=\"…\"` 属性、(4) HTML の**取得系属性**(`<link href>` `<script src>` `<img src|srcset>` `<iframe src>` `<object data>` SVG の `href`/`xlink:href` 等。`<a href>` は組版中に取得を起こさないので対象外)、(zip 経路は加えて展開した `.css` / `.html` 全件)。拒否するのは **許可リスト外の at-rule**(`@import` `@use` 等。`@media` `@page` とマージンボックス `@bottom-center` 等は許可)と、**外部を指す URL 値**(scheme 付き・scheme 相対 `//host/x`・許可外の `data:`)で、`url()` に限らず`image-set(\"http://…\")` のような引用符文字列も含む。許可する `data:` は `data:image/png` `data:image/jpeg` `data:image/jpg` `data:image/gif` `data:image/webp` `data:font/` `data:application/font-woff` の接頭辞のみ(`data:image/svg+xml` は不可)。**相対 URL と断片(`#id`)は通る。むしろ必須である** — 文書は作業フォルダの `doc/` に置かれたものとして扱い、テンプレ CSS・共通フォント・テンプレ JS・画像を `../css/…` `../css/fonts/…` `../js/…` `../images/…` の相対パスで参照する。サーバは参照された実体だけを `doc/` の兄弟(`css/` `js/` `images/`)へ同梱する。文書直下基準の `css/…` `images/…` は `doc/css/…` を指し、同梱されない。画像は `../images/<名前>` と `../images/<フォルダ>/<名前>`(1 段まで)を同梱し、2 段以上は同梱しない。リクエストの `css` は `css/<会社>_<ファンド>_<版種>.css` の位置に置かれた CSS として解釈する(相対 `url()` は CSS 自身の位置が基準。例: `url(fonts/a.woff2)` → `css/fonts/a.woff2`、`url(../images/a.svg)` → `images/a.svg`)。文書へは `doc/` から見た形に付け替えて埋め込む。相対参照は `url()` で書くこと(引用符文字列の相対参照は解決されない)。同梱の実体が無い相対参照の `<link>` / `<script src>` は 400 にはせず要素ごと落とす(404 は組版のページ分割を止めるため)。**削らずに拒む**のは外部参照の方で、違反が 1 件でも PDF は生成されない。また、タグ境界が一意に決まらない HTML(閉じないタグ・コメント・`<style>`/`<script>`)は`code=DOCUMENT_UNPARSABLE` の 400 で拒む — 検査できない入力を通すとそれ自体が回避路になる。\n\n**文書内の JavaScript**: 組版時に実行される。組版エンジンは文書を再パースして script をビューアの window へ作り直すため、`<head>` で `DOMContentLoaded` に登録した処理は発火しない(実測。処理は body 末尾に置くこと)。`<script src=\"../js/x.js\">` は、サーバが同梱した実体の中身をその位置へ**インライン展開**してから組版へ渡すので実行される(そのままでは相対 URL の解決基準がずれて 404 になるため)。同梱の実体が無い参照は要素ごと落ち、2MB を超えるファイルは展開されず 404 で実行されない。なお組版ブラウザの **HTTP/HTTPS 通信**は、そのビルド専用の loopback オリジン 1 つだけへ中継される(それ以外は宛先が loopback でも 502 で落ちる)。ただし遮断の実体は HTTP プロキシ 1 本なので、**HTTP 以外の経路**(WebRTC の UDP 等)はこの中継を通らない — そこは残余リスクとして残る。\n\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildMergeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PDF バイナリ",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "description": "PDF バイナリ"
                }
              }
            }
          },
          "400": {
            "description": "リクエスト検証エラー (kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "413": {
            "description": "リクエストが大きすぎる (kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "500": {
            "description": "サーバ内部エラー (kind=unexpected)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
    "/preview": {
      "get": {
        "tags": [
          "vivliostyle"
        ],
        "summary": "稼働中のプレビューセッション一覧(自分が作成した分のみ)",
        "operationId": "listPreviews",
        "description": "セッションは作成した利用者に紐づく。admin 以外は自分が作成したセッションだけが返る。",
        "responses": {
          "200": {
            "description": "セッション一覧",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewSessionList"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "vivliostyle"
        ],
        "summary": "ライブプレビューを起動 (inline JSON または project zip)",
        "operationId": "startPreview",
        "description": "inline は `application/json` (BuildInlineRequest)、project は `application/zip`。返却 `url` (`/api/preview/{id}/`) を同一オリジンで開くと vivliostyle ビューアが表示される。\n\nzip 経路の受入条件と任意クエリ(`entry` `size` `singleDoc`)は `POST /build/project` と**同一**である(同じ展開・検証を通る):\n\nリクエストボディに project zip を `application/zip` で送信する。**受理は許可リスト方式**で、1 エントリでも条件を外れた時点で zip 全体が 400 になる(該当エントリだけ無視する挙動ではない)。\n\n**展開して残せる拡張子**(小文字化した basename の末尾一致): `.html` `.htm` `.xhtml` `.xht` `.md` `.markdown` `.css` `.css.map` `.png` `.jpg` `.jpeg` `.svg` `.gif` `.webp` `.apng` `.ttf` `.otf` `.woff` `.woff2` `.json`。これ以外(`.js` `.mjs` `.cjs` `.ts` を含む)は 400。設定を `vivliostyle.config.js` 等で送っても「無視される」のではなく**拒否される**。\n\n**黙って捨てる OS ノイズ**(400 にはしない): `__MACOSX/**`・`.DS_Store`・`Thumbs.db`・`desktop.ini`・`._` で始まる名前。\n\n**拒否する npm 成果物**: `node_modules/` を含むパス・`package.json`・`package-lock.json`・`.npmrc`。\n\n**config は `vivliostyle.config.json` ちょうど 1 件**(名前は大小文字を無視して照合)。2 件以上あれば 400。無ければ単一エントリを既定として使う。内容は**厳密な JSON** のみで、コメント・末尾カンマ(JSONC)や配列(複数タスク)は 400。\n\n**受理するトップレベルキー**: `entry` `entryContext` `theme` `cover` `toc` `title` `author` `language` `readingProgression` `size` `copyAsset`。これ以外はキー名を挙げて 400 になる(`output` `workspaceDir` `vfm` `image` `timeout` `viewer` `server` `static` `pdfPostprocess` など vivliostyle CLI の正規フィールドも含む)。`base` は専用メッセージで 400 — 文書の配信先はサーバが固定する。\n\n**`theme` の制約**: 展開ルート内に実在する `.css` ファイルへの相対パスのみ。拡張子の判定は**大小文字を区別**する(`.CSS` は不可)。npm パッケージ指定子・`{specifier, import}` 形は不可。`entry` 要素ごとの `theme` にも同じ制約が掛かる。\n\n**400 応答**はすべて `kind=validation` で、`message` は日本語 1 文。拡張子違反は「取り込めるのは <許可拡張子一覧> だけです」、キー違反は「vivliostyle.config.json に指定できないキーがあります: <キー名>」のように**違反した対象を名指しする**(どのファイル・どのキーで落ちたかが応答から判る)。展開時の資源上限(ファイル数・展開後サイズ・圧縮比・目録サイズ)超過も 400 で、上限値を message に含める。413 になるのは**リクエスト本文そのもの**がサーバの `bodyLimit` を超えた場合だけ(fastify が返す)。\n\n",
        "responses": {
          "201": {
            "description": "起動したセッション",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewSession"
                }
              }
            }
          },
          "400": {
            "description": "リクエスト検証エラー (kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "403": {
            "description": "編集者権限が必要 (viewer は不可。kind=forbidden)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "413": {
            "description": "プロジェクト zip が大きすぎる (zip 経路のみ。kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "500": {
            "description": "サーバ内部エラー (kind=unexpected)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
    "/preview/{id}": {
      "get": {
        "tags": [
          "vivliostyle"
        ],
        "summary": "プレビューセッションのメタを取得",
        "operationId": "getPreview",
        "description": "他人が作成したセッションは 404(存在を漏らさないため 403 にしない)。",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "セッション",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewSession"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "404": {
            "description": "対象が存在しない (kind=not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "vivliostyle"
        ],
        "summary": "プレビューセッションを停止",
        "operationId": "stopPreview",
        "description": "他人が作成したセッションは 404(停止も作業ディレクトリ削除も起きない)。",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "停止完了"
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "403": {
            "description": "編集者権限が必要 (viewer は不可。kind=forbidden)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "404": {
            "description": "対象が存在しない (kind=not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuildMergeRequest": {
        "type": "object",
        "properties": {
          "documents": {
            "minItems": 1,
            "maxItems": 30,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BuildMergeDocument"
            },
            "description": "結合する文書。配列順に連結し通しページ番号を振る"
          },
          "size": {
            "description": "ページサイズ (既定 A4)",
            "example": "A4",
            "type": "string"
          }
        },
        "required": [
          "documents"
        ]
      },
      "BuildMergeDocument": {
        "type": "object",
        "properties": {
          "html": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4194304,
            "description": "レンダリング済み(nunjucks)HTML"
          },
          "css": {
            "default": "",
            "type": "string",
            "maxLength": 1048576
          }
        },
        "required": [
          "html"
        ]
      },
...
      "AppError": {
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/AppErrorKind"
          },
          "message": {
            "type": "string",
            "description": "ユーザ向け(JP)。常に表示して安全"
          },
          "code": {
            "description": "機械可読コード 例: 'USER_DISABLED'",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "message"
        ],
        "additionalProperties": false
      },
      "AppErrorKind": {
        "type": "string",
        "enum": [
          "not_found",
          "validation",
          "unauthorized",
          "forbidden",
          "conflict",
          "network",
          "unexpected"
        ]
      },
...
      "PreviewSessionList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PreviewSession"
        }
      },
      "PreviewSession": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "プレビューセッション ID"
          },
          "mode": {
            "type": "string",
            "enum": [
              "inline",
              "project"
            ]
          },
          "createdAt": {
            "type": "string",
            "description": "作成時刻 (ISO)"
          },
          "expiresAt": {
            "type": "string",
            "description": "アイドル失効予定時刻 (ISO)"
          },
          "url": {
            "type": "string",
            "description": "同一オリジンのプレビュー URL",
            "example": "/api/preview/{id}/"
          }
        },
        "required": [
          "id",
          "mode",
          "createdAt",
          "expiresAt",
          "url"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "editor.sid",
        "description": "ログイン時に発行されるセッション cookie。"
      }
    }
  }
}